ring: fix BLE stability, correct protocol decoding, add sleep insights
BLE Connection Fixes (RingBLEClient):
- Don't close GATT on passive disconnect (kills autoConnect)
- removeBond() only on Forget, not on disconnect
- Keepalive switched from 0x0C to 0x3A (official SDK ping)

Protocol Corrections (from APK decompilation + Gadgetbridge):
- 0x23 = combined BP measurement (not SpO2)
- 0x24 = combined result: HR + systolic + diastolic + SpO2 + stress
- SpO2-only uses 0x3E/0x3F
- 0x02 = user profile (not activity query)
- 0x20 = device capabilities (not static config)
- 28+ commands renamed to match KeepFit SDK constants

Decoder Rewrite (RingDecoder.kt):
- 0x24: extract all 5 metrics (was only SpO2)
- 0x14: use ring's timestamp, detect measurement errors
- 0x0B: capture charging state
- 0x16: multi-packet HR history with 6-sample averaging
- 0x10: 15x 1-min activity buckets per packet
- 0x3F: standalone SpO2 result

Sleep Detection:
- Threshold-based stage decoding (>=80 deep, >=1 light)
- History query sends day parameter (was 0 = no-op)
- Stage blocks persisted with run-length encoding
- Real sleep score in UI (was totalMinutes/5)
- Coach insight card with personalized recommendations

Documentation:
- keepfit-protocol-complete.md (APK decomp dump)
- per-stat-decoding.md (byte-level reference)
- ring-protocol.md (authoritative protocol spec)
- protocol-discoveries.md (corrected assumptions)
- development-setup.md (Java + build instructions)

+++ android/app/src/main/java/com/pulseloop/ring/ColmiDecoder.kt
@@ -47,8 +47,8 @@ object ColmiDecoder {
         ColmiCommandID.NOTIF_BATTERY -> listOf(RingDecodedEvent.Battery(percent = v[2].toInt()))
         ColmiCommandID.NOTIF_LIVE_ACTIVITY -> {
             val steps = ColmiBytes.u24(v[2], v[3], v[4])
-            val calories = ColmiBytes.u24(v[5], v[6], v[7]).toDouble() / 10.0
-            val distance = ColmiBytes.u24(v[8], v[9], v[10]).toDouble()
+            val calories = ColmiBytes.u24(v[5], v[6], v[7]).toInt() / 10
+            val distance = ColmiBytes.u24(v[8], v[9], v[10]).toInt()
             listOf(RingDecodedEvent.ActivityUpdate(
                 _timestamp = now, steps = steps, distanceMeters = distance, calories = calories
             ))
@@ -142,7 +142,7 @@ object ColmiDecoder {
         if (ts < lower || ts > upper) return emptyList()
 
         val steps = ColmiBytes.u16(v[9], v[10])
-        val distance = ColmiBytes.u16(v[11], v[12]).toDouble()
+        val distance = ColmiBytes.u16(v[11], v[12]).toInt()
         return listOf(RingDecodedEvent.ActivityBucket(
             _timestamp = ts, steps = steps, distanceMeters = distance
         ))
